xero/code/requirements.txt
apache-airflow==2.3.0
apache-beam==2.38.0
google-cloud-storage==2.0.0
google-cloud-secret-manager==2.12.0
oauthlib==3.2.0
requests-oauthlib==1.3.1
cachetools==5.0.0
structlog==22.1.0
flask==2.1.0